requirement.txt
flask
mysql-connector-python
google-generativeai
python-dotenv